move Ioc to Prism core
