+++ Source/Prism/Ioc/IContainerRegistry.cs
